Fixed various code smells

+++ src/pipes/evaluator_pipe.cpp
@@ -17,11 +17,15 @@ double EvaluatorPipe::evaluate(const std::vector<unsigned char>& program_data) {
   return evaluator_.evaluate(session);
 }
 
-uint32_t EvaluatorPipe::getMemorySize() const { return variable_count_; }
+uint32_t EvaluatorPipe::getMemorySize() const { return static_cast<uint32_t>(variable_count_); }
 
-uint32_t EvaluatorPipe::getStringTableSize() const { return string_table_count_; }
+uint32_t EvaluatorPipe::getStringTableSize() const {
+  return static_cast<uint32_t>(string_table_count_);
+}
 
-uint32_t EvaluatorPipe::getStringTableItemLength() const { return max_string_size_; }
+uint32_t EvaluatorPipe::getStringTableItemLength() const {
+  return static_cast<uint32_t>(max_string_size_);
+}
 
 const std::vector<AggregationEvaluator::EvaluatorDescription>&
 EvaluatorPipe::getEvaluators() const {
